+++ research/conditional/moe_layers/expert_types.py
@@ -27,12 +27,15 @@ class ExpertFF(LoggingLayer):
         self.doutput = dmodel if doutput is None else doutput
         self.n_experts = n_experts
         self.use_einsum = use_einsum
+        self.expert_size = expert_size
         self.activation = resolve_activation_name(activation_name)
 
         fan_in_factor = topk if use_topk_initialization else n_experts
-        init = get_init_fun(init_type=init_type, init_scale=init_scale)
-        self.lin1_weight = init(shape=(n_experts, dmodel, expert_size), fan_in=dmodel)
-        self.lin2_weight = init(
+        self.init_fun = get_init_fun(init_type=init_type, init_scale=init_scale)
+        self.lin1_weight = self.init_fun(
+            shape=(n_experts, dmodel, expert_size), fan_in=dmodel
+        )
+        self.lin2_weight = self.init_fun(
             shape=(n_experts, expert_size, self.doutput),
             fan_in=int(fan_in_factor * expert_size),
         )
@@ -72,31 +75,14 @@ class ExpertFF(LoggingLayer):
 class ExpertGated(ExpertFF):
     def __init__(
         self,
-        dmodel: int,
-        n_experts: int,
-        expert_size: int,
-        init_type: str,
-        init_scale: float,
-        use_einsum: bool = False,
-        doutput: Optional[int] = None,
+        *args,
         activation_name: str = "silu",
-        topk: int = 1,
-        use_topk_initialization: bool = False,
+        **kwargs,
     ):
-        super().__init__(
-            dmodel,
-            n_experts,
-            expert_size,
-            init_type,
-            init_scale,
-            use_einsum,
-            doutput,
-            activation_name,
-            topk,
-            use_topk_initialization,
+        super().__init__(*args, activation_name=activation_name, **kwargs)
+        self.gate_weight = self.init_fun(
+            shape=(self.n_experts, self.dmodel, self.expert_size), fan_in=self.dmodel
         )
-        init = get_init_fun(init_type=init_type, init_scale=init_scale)
-        self.gate_weight = init(shape=(n_experts, dmodel, expert_size), fan_in=dmodel)
 
     @time_measured("process_by_experts")
     def forward(self, x: torch.Tensor):
